refactor: CartItem 에서 Quantity 추가
