+++ src/main/java/cart/dao/CartItemDao.java
@@ -1,17 +1,18 @@
 package cart.dao;
 
-import cart.domain.CartItem;
 import cart.domain.Member;
-import cart.domain.Product;
+import cart.domain.Quantity;
+import cart.domain.product.CartItem;
+import cart.domain.product.Product;
+import java.sql.PreparedStatement;
+import java.sql.Statement;
+import java.util.Collections;
+import java.util.List;
+import java.util.Objects;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.support.GeneratedKeyHolder;
 import org.springframework.jdbc.support.KeyHolder;
 import org.springframework.stereotype.Repository;
-
-import java.sql.PreparedStatement;
-import java.sql.Statement;
-import java.util.List;
-import java.util.Objects;
 
 @Repository
 public class CartItemDao {
@@ -22,11 +23,13 @@ public class CartItemDao {
     }
 
     public List<CartItem> findByMemberId(Long memberId) {
-        String sql = "SELECT cart_item.id, cart_item.member_id, member.email, product.id, product.name, product.price, product.image_url, cart_item.quantity " +
-                "FROM cart_item " +
-                "INNER JOIN member ON cart_item.member_id = member.id " +
-                "INNER JOIN product ON cart_item.product_id = product.id " +
-                "WHERE cart_item.member_id = ?";
+        String sql =
+                "SELECT cart_item.id, cart_item.member_id, member.email, product.id, product.name, product.price, product.image_url, cart_item.quantity "
+                        +
+                        "FROM cart_item " +
+                        "INNER JOIN member ON cart_item.member_id = member.id " +
+                        "INNER JOIN product ON cart_item.product_id = product.id " +
+                        "WHERE cart_item.member_id = ?";
         return jdbcTemplate.query(sql, new Object[]{memberId}, (rs, rowNum) -> {
             String email = rs.getString("email");
             Long productId = rs.getLong("product.id");
@@ -35,9 +38,10 @@ public class CartItemDao {
             String imageUrl = rs.getString("image_url");
             Long cartItemId = rs.getLong("cart_item.id");
             int quantity = rs.getInt("cart_item.quantity");
-            Member member = new Member(memberId, email, null);
+            //TODO: 쿠폰도 조회할 수 있도록 변경해야 함
+            Member member = new Member(memberId, email, null, Collections.emptyList());
             Product product = new Product(productId, name, price, imageUrl);
-            return new CartItem(cartItemId, quantity, product, member);
+            return new CartItem(cartItemId, new Quantity(quantity), product, member);
         });
     }
 
@@ -52,7 +56,7 @@ public class CartItemDao {
 
             ps.setLong(1, cartItem.getMember().getId());
             ps.setLong(2, cartItem.getProduct().getId());
-            ps.setInt(3, cartItem.getQuantity());
+            ps.setInt(3, cartItem.getQuantity().getValue());
 
             return ps;
         }, keyHolder);
@@ -61,11 +65,13 @@ public class CartItemDao {
     }
 
     public CartItem findById(Long id) {
-        String sql = "SELECT cart_item.id, cart_item.member_id, member.email, product.id, product.name, product.price, product.image_url, cart_item.quantity " +
-                "FROM cart_item " +
-                "INNER JOIN member ON cart_item.member_id = member.id " +
-                "INNER JOIN product ON cart_item.product_id = product.id " +
-                "WHERE cart_item.id = ?";
+        String sql =
+                "SELECT cart_item.id, cart_item.member_id, member.email, product.id, product.name, product.price, product.image_url, cart_item.quantity "
+                        +
+                        "FROM cart_item " +
+                        "INNER JOIN member ON cart_item.member_id = member.id " +
+                        "INNER JOIN product ON cart_item.product_id = product.id " +
+                        "WHERE cart_item.id = ?";
         List<CartItem> cartItems = jdbcTemplate.query(sql, new Object[]{id}, (rs, rowNum) -> {
             Long memberId = rs.getLong("member_id");
             String email = rs.getString("email");
@@ -75,9 +81,9 @@ public class CartItemDao {
             String imageUrl = rs.getString("image_url");
             Long cartItemId = rs.getLong("cart_item.id");
             int quantity = rs.getInt("cart_item.quantity");
-            Member member = new Member(memberId, email, null);
+            Member member = new Member(memberId, email, null, Collections.emptyList());
             Product product = new Product(productId, name, price, imageUrl);
-            return new CartItem(cartItemId, quantity, product, member);
+            return new CartItem(cartItemId, new Quantity(quantity), product, member);
         });
         return cartItems.isEmpty() ? null : cartItems.get(0);
     }
