+++ src/main/java/cart/dto/CartItemResponse.java
@@ -1,6 +1,6 @@
 package cart.dto;
 
-import cart.domain.CartItem;
+import cart.domain.product.CartItem;
 
 public class CartItemResponse {
 
@@ -20,7 +20,7 @@ public class CartItemResponse {
     public static CartItemResponse of(CartItem cartItem) {
         return new CartItemResponse(
                 cartItem.getId(),
-                cartItem.getQuantity(),
+                cartItem.getQuantity().getValue(),
                 ProductResponse.of(cartItem.getProduct())
         );
     }
